cleanup after previous stuff; 0.25 speed still
There is a really weird problem going on here. When we apply a voltage
(like 12.0 volts), and we ask the motor for its voltage, it reads
exactly a quarter of that. What's up with this?

We have determined that mass doesn't matter - the elevator simulation
never interacts with the motor. I'm not super sure what's going on; I
hope that I can figure it out, or it's just a sim thing.