timeout fixes + helpful shell utility
The phoenix docs changed, so we shifted from
Utils.getCurrentTimeSeconds(), but rather now we use
Timer.getFPGATimestamp()

This seems in more reasonable in sim.

Also, we added a utility shell script to the repo base that has the
ability to extract (and then remove) log files from the roborio -
including vision logs.