improvements to auton
Firstly, the default command that is supposed to run in auton wasn't
working properly. This is because auton runs on one big sequential
command group, which holds the requirements of all of its subcommands,
recusively. This means that when we had any single command that controls
elevator (like score), elevator was effectively on reserve for the
entire auton, i.e. default command never ran. Turning all
elevator-related commands into proxy commands will fix this.

Secondly, the alignment for intake lineup should end as soon as a piece
is present. A small error in vision that puts the setpoint outside of
intake bounds would previously have blocked auton from doing anything
else. Now it ends when the intakeReady condition is met. This also means
that after intaking, we go into the next path right away.