proxied commands nearly killed the bot... we should fix that
For now, I've removed all of those commands. This means that
elevator-prefire is no longer a thing, but we can hope to work that out
later.

The cause of this problem seems to be that the proxy commands were not
JUST elevator commands, they included other subsystems too (namely coral
intake). When the proxied command asked for requirements on the coral
intake, it killed the other holder of the coral intake commands: the
whole command group that the PROXIED COMMAND ITSELF lived in!

This caused the proxied command to self-destruct and move to the next
phase without actually doing anything. This caused the scoring sequence
to end (because it uses coral rollers).

To fix this issue, we should avoid making mega-commands that use tons
of subsystems, and instead break the comands into smaller commands that
operate on a per-subsystem basis (LIKE IVE BEEN SAYING ALL ALONG). This
change would make sure that we aren't accidentally killing commands that
are "important", like scoring all the way.