rotation was the failure, I guess.
when the drivetrain is rotating, it cannot as easily drive in the right
direction, so the path started funny. When it worked well, the robot was
already in the correct orientation (the later half of the path).
Although the non-instant change in velocity is not great, it is MUCH
less problematic than I had previously believed.

also removed ideal path to save compute time.