Initial design
Basically, theres a base class, LEDState, which all states must extend.
An instance of each state should live inside the LEDs class on
construction to avoid overuse of memory.

Each periodic loop, the correct state is chosen and applied.

The LEDs class provides the logic to switch between states, but the
states themselves contain the actual ControlRequest to be set. This
means that modifying the behavior of a state doesn't affect the LED
class. Changing when a state is active, however, does.