Implement coral lights and bad controller warning
turns out that a GenericHIDController just has a method isConneted() so
we don't need to string searches or whatever to tell if a controller is
connected.

I think this is better because it's a little less work AND because we
have regular controllers and backup controllers, we don't have to worry
about changing names.