Modify Superstructure.enter() to always proxy its commands.
This means that we can omit the .asProxy() whenever we need a
subsystem's passive behavior to run in Auton. This also means that we
never have to worry about proxied commands killing themselves, because
all commands are proxied now.

This also fixes the issue that can occur in Auton where the robot waits
to intake without coral present.