Break vision and apriltag logic down into separate sections
This breaks up simple stuff, like the IO wrapper for cameras, and
apriltag-related stuff.

This should help to avoid redundant code, or a super large vision/
folder.

All vision that has to do with localization is in a localization
package, and anything that uses april tags has AprilTag in the name.
This creates large-ish names, so formatting across the project may need
to be remedied.